+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Add support for October CMS 4.x.
+  * Minimal version of PHP required is 8.2.0
+* Drop support for October CMS 3.x.
+
 ## [3.1.11] - 2023-11-30
 
 * Add option to keep query string when redirecting.